backend/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
flask-cors
marshmallow-sqlalchemy
marshmallow-enum
sqlalchemy
flask-jwt-extended
flask-bcrypt
requests
